finished moving error checking to presenters. Code Done

+++ Software_Eng_ASOEE_2024/app/src/main/java/com/example/software_eng_asoee_2024/views/ActiveSubstanceEdit/Deletion/ActiveSubstanceDeletePresenter.java
@@ -20,8 +20,13 @@ public class ActiveSubstanceDeletePresenter {
     }
 
     public boolean deleteActiveSubstance(ActiveSubstance ac) {
+        if(ac == null) {
+            view.showMessage("None selected to be edited");
+            return false;
+        }
         this.activeSubstanceDAO.delete(ac);
         createActiveSubstanceSpinner();
+        view.showMessage("Done!");
         return(this.activeSubstanceDAO.findAll().isEmpty());
     }
