+++ Software_Eng_ASOEE_2024/app/src/main/java/com/example/software_eng_asoee_2024/views/PharmaceuticalProductEdit/Deletion/PharmaceuticalProductDeletePresenter.java
@@ -20,8 +20,13 @@ public class PharmaceuticalProductDeletePresenter {
     }
 
     public boolean deletePharmaceuticalProduct(PharmaceuticalProduct ac) {
+        if(ac == null) {
+            view.showMessage("None selected to be delete");
+            return false;
+        }
         this.pharmaceuticalProductDAO.delete(ac);
         createPharmaceuticalProductSpinner();
+        view.showMessage("Done!");
         return(this.pharmaceuticalProductDAO.findAll().isEmpty());
     }
